test(api): split sim.test.ts by concern
Feedback item 4. The file had grown to ~1450 lines, the largest in the repo.
Split on its existing describe boundaries into gating, clock, scenarios,
fixtures, reset, and replay, with the cross-describe helpers moved to
test/setup/sim-helpers.ts. Largest resulting file is 312 lines.

No test was removed: a review pass established there are no redundant ones —
the two apparent overlaps (reset asserted two ways, clock projection
asserted through ingestion and through the fixtures endpoint) each prove a
different code path. Titles and bodies are unchanged; only imports moved.
301 integration tests before and after.

Helpers used by a single file stayed in it rather than inflating the shared
surface. `fileParallelism: false` means the split costs no wall-clock time.
